Fix occasional long transition out of Cutscenes.
The cause: bgm_measure_f (and related variables) are not guaranteed to
hit every single frame value; every once in a while, frames are skipped
or repeated due to an unusually short or long sim frame. The previous
code only advanced if bgm_measure_f was exactly BGM_FRAMES_PER_MEASURE,
when sometimes it would go 54,54,1 (causing the loop to repeat for four
extra measures.

bgm_measure_ff is guaranteed to be true on the first frame of a new measure,
even if bgm_measure_f is 1 (i.e. if frame 0 got skipped). This will
occasionally cause a slight hiccup in transitions, but it's better than
four extra measures.